met-cron/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest
pytest-mock
requests
pyhamcrest
pytest-cov

# Lint and code style
flake8==7.0.0
flake8-blind-except
flake8-debugger
flake8-docstrings
flake8-isort
flake8-quotes
pep8-naming
autopep8
coverage
pylint==3.1.0
pylint-flask
FreezeGun

# docker
lovely-pytest-docker
pytest-asyncio
